basic finished

+++ code/module/minimap.py
@@ -3,20 +3,28 @@ import subprocess
 
 from config import config
 from prototype.module import Module
-from moduleConfig.minimapConfig import minimapConfig
+from moduleConfig.minimapConfig import MinimapConfig
 from moduleResult.minimapResult import MinimapResult
 from moduleResult.alignment import Alignment
 
+from utils import IOUtils
+
 class Minimap(Module):
-    def __init__(self):
-        super().__init__()
+    def __init__(self, config:MinimapConfig):
+        super().__init__(config)
+        self.minimapConfig = config
 
     def run(self):
         queryFile = f"{config.datasetBase}/{config.datasetName}.fasta"
-        resultFolder = f"{config.resultBase}/minimapResult-{minimapConfig.name}"
+        resultFolder = f"{config.resultBase}/minimapResult-{self.minimapConfig.name}"
 
+        if (os.path.exists(resultFolder)):
+            IOUtils.showInfo(f'Skipped minimap on {config.datasetName}')
+            return
+        
         os.makedirs(resultFolder)
-        command = minimapConfig.getMinimapCommand(queryFile)
+        IOUtils.showInfo(f"Begin minimap on {config.datasetName}")
+        command = self.minimapConfig.getMinimapCommand(queryFile)
         with open(f"{config.tempFolder}/alignment.sam", 'wt') as fp:
             subprocess.run(command, shell=True, stdout=fp, stderr=subprocess.DEVNULL)
         with open(f"{config.tempFolder}/alignment.sam") as fp:
@@ -26,24 +34,20 @@ class Minimap(Module):
                     f.write(line)
         os.remove(f"{config.tempFolder}/alignment.sam")
     
-    def getResults(self, sampleList):
-        resultFolder = f"{config.resultBase}/minimapResult-{minimapConfig.name}"
-        if (not os.path.exists(resultFolder)):
-            self.run()
+    def getResult(self, sample):
+        resultFolder = f"{config.resultBase}/minimapResult-{self.minimapConfig.name}"
         
-        for sample in sampleList:
-            if (os.path.exists(f"{resultFolder}/{sample.query}.sam")):
-                result = MinimapResult()
-                with open(f"{resultFolder}/{sample.query}.sam") as fp:
-                    reads = fp.readlines()
-                for read in reads:
-                    terms = read.strip().split('\t')
-                    referenceName = terms[2]
-                    CIGAR = terms[5]
-                    mappingQuality = int(terms[4])
-                    if (referenceName != '*'):
-                        result.addAlignment(Alignment(referenceName, mappingQuality, CIGAR))
-                if (result.bestAlignment is not None):
-                    sample.addResult("minimap", result)
-
-minimap = Minimap()
+        result = MinimapResult()
+        if (os.path.exists(f"{resultFolder}/{sample.query}.sam")):
+            with open(f"{resultFolder}/{sample.query}.sam") as fp:
+                reads = fp.readlines()
+            for read in reads:
+                terms = read.strip().split('\t')
+                referenceName = terms[2]
+                CIGAR = terms[5]
+                mappingQuality = int(terms[4])
+                if (referenceName != '*'):
+                    result.addAlignment(Alignment(referenceName, mappingQuality, CIGAR))
+        if (result.bestAlignment is None):
+            result = None
+        return result
